Remove some obsolete code for XADataSource handling
